requirements.txt
jinja2==2.10.1
click==7.0
gitpython==3.1.2
virtualenv==20.0.20
